+++ UiPathOrch/OrchestratorCmdlet/FolderUserCmdlet/CopyFolderUser.cs
@@ -65,7 +65,7 @@ public class CopyFolderUserCmdlet : OrchestratorPSCmdlet
             var results = ParallelResults.GroupBy(drivesFolders, df =>
             {
                 return IncludeInherited
-                    ? df.drive.FolderUsersWithInherited.Get(df.folder)
+                    ? df.drive.GetFolderUsersUnion(df.folder)
                     : df.drive.FolderUsersWithNoInherited.Get(df.folder);
             });